benchmarks/requirements.txt
Django
ipython
requests
django-debug-toolbar
redis
hiredis
python-dotenv

tuna
